refactor:add mobile styles for signIn and authCard
- Update import paths
- Update lib version to 0.0.8
- Add index.ts file to config folder
- Update styles for signUp and layout
-
