+++ src/features/auth/ui/signIn/signIn.module.scss
@@ -1,11 +1,14 @@
-/* stylelint-disable-next-line scss/double-slash-comment-whitespace-inside */
-//TODO: make styles for mobile
+@use '@/shared/styles/mixins' as *;
 
 .form {
   display: flex;
   flex-direction: column;
   gap: 30px;
   margin-bottom: 18px;
+
+  @include mobile {
+    gap: 5px;
+  }
 
   & .container {
     display: flex;
@@ -22,6 +25,10 @@
     display: flex;
     flex-direction: column;
     gap: var(--auth-form-gap);
+
+    @include mobile {
+      gap: 65px;
+    }
 
     & .password {
       cursor: pointer;
