+++ src/features/auth/ui/signUp/signUp.module.scss
@@ -8,7 +8,7 @@
 
   width: 100%;
   max-width: 378px;
-  height: auto;
+  height: 100%;
   padding: 25px;
 }
