+++ src/features/layout/layout.module.scss
@@ -1,5 +1,13 @@
+@use '@/shared/styles/mixins' as *;
+
 .authLayout {
   display: flex;
   align-items: center;
   justify-content: center;
+  height: calc(100vh - var(--header-height));
+
+  @include mobile {
+    align-items: flex-start;
+    justify-content: flex-start;
+  }
 }
